+++ user-operator/src/main/java/io/strimzi/operator/user/model/KafkaUserModel.java
@@ -9,6 +9,7 @@ import io.fabric8.kubernetes.api.model.OwnerReference;
 import io.fabric8.kubernetes.api.model.OwnerReferenceBuilder;
 import io.fabric8.kubernetes.api.model.Secret;
 import io.fabric8.kubernetes.api.model.SecretBuilder;
+import io.strimzi.api.ResourceAnnotations;
 import io.strimzi.api.kafka.model.user.KafkaUser;
 import io.strimzi.api.kafka.model.user.KafkaUserAuthentication;
 import io.strimzi.api.kafka.model.user.KafkaUserAuthorizationSimple;
@@ -20,6 +21,7 @@ import io.strimzi.api.kafka.model.user.acl.AclRule;
 import io.strimzi.certs.CertAndKey;
 import io.strimzi.certs.CertManager;
 import io.strimzi.certs.OpenSslCertManager;
+import io.strimzi.operator.common.Annotations;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.ReconciliationLogger;
 import io.strimzi.operator.common.Util;
@@ -244,7 +246,11 @@ public class KafkaUserModel {
             String userCrt = userSecret.getData().get("user.crt");
             String userKey = userSecret.getData().get("user.key");
 
-            if (originalCaCrt != null
+            if (userSecret.getMetadata() != null
+                    && Annotations.booleanAnnotation(userSecret, ResourceAnnotations.ANNO_STRIMZI_IO_FORCE_RENEW, false)) {
+                // The user secret has the annotation which forces replacement => we have to generate a new user certificate
+                this.userCertAndKey = generateNewCertificate(reconciliation, clientsCa);
+            } else if (originalCaCrt != null
                     && originalCaCrt.equals(caCrt)
                     && userCrt != null
                     && !userCrt.isEmpty()